% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmiss-internal-sim-study.R
\name{evmiss-internal-sim-study}
\alias{evmiss-internal-sim-study}
\alias{rl_sim_study}
\alias{rfrechet}
\alias{rmaxAR}
\alias{MCSEbias}
\alias{MCSEsd}
\alias{MCSErmse}
\alias{MCSEstatistics}
\alias{MCSEstatistics_boot}
\alias{MCSEstatistics_boot2}
\alias{MCSEstatistics_boot3}
\title{Internal evmiss simulation study functions}
\usage{
rl_sim_study(
  object,
  return_periods,
  level,
  profile,
  mult,
  faster,
  epsilon,
  timeout
)

rfrechet(n)

rmaxAR(n, theta, scale = 1, shape = 0)

MCSEbias(x, truth = 0)

MCSEsd(x, truth = 0)

MCSErmse(x, truth = 0)

MCSEstatistics(x, truth = 0)

MCSEstatistics_boot(x, B = 100, truth = 0)

MCSEstatistics_boot2(x, B = 100, truth = 0)

MCSEstatistics_boot3(x, B = 100, truth = 0)
}
\description{
Internal evmiss simulation study functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
