% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_evmiss_sim_study_2.R
\name{sim_study_2_methods}
\alias{sim_study_2_methods}
\alias{plot.evmiss_sim_study_2}
\alias{summary.evmiss_sim_study_2}
\alias{print.summary.evmiss_sim_study_2}
\alias{tab.evmiss_sim_study_2}
\title{Methods for objects of class \code{"evmiss_sim_study_2"}}
\usage{
\method{plot}{evmiss_sim_study_2}(
  x,
  what = c("return", "mu", "sigma", "xi", "all"),
  distn = c("marginal", "joint"),
  approach = c("full", "adjust", "naive", "discard", "weight1", "weight2"),
  return_period = 100,
  vertical = TRUE,
  main = c("full", "adjust", "naive", "discard", "weight1", "weight2"),
  penultimate = TRUE,
  line_col = c("orange", "purple", "red", "blue"),
  line_lty = c(1, 1, 2, 3),
  legend = 4,
  mar = c(5, 5, 2, 2),
  layout = FALSE,
  no_ylab = FALSE,
  ...
)

\method{summary}{evmiss_sim_study_2}(
  object,
  ...,
  what = c("all", "return"),
  return_period,
  statistics,
  meanSE = FALSE,
  log = FALSE,
  log_return = FALSE,
  vsfull = TRUE
)

\method{print}{summary.evmiss_sim_study_2}(x, quiet = FALSE, ...)

\method{tab}{evmiss_sim_study_2}(object, return_period, ...)
}
\arguments{
\item{x, object}{An object inheriting from class \code{"evmiss_sim_study_2"}, a
result of a call to the function \code{\link{sim_study_2}}. For
\code{plot.evmiss_sim_study_2}, further arguments to be passed to
\link[graphics:plot.default]{plot.default} and \code{\link[graphics:abline]{abline}},
such as graphical parameters.}

\item{what}{A character scalar. What should be plotted? See \code{distn} for
further details. The default is \code{"return"} for the \code{plot} method and
\code{"all"} for the \code{summary} method.
\itemize{
\item \code{"return"}: estimates of return levels using the approaches \code{"full"},
\code{"adjust"}, \code{"naive"} and \code{"discard"}.
\item \code{"mu"} , \code{"sigma"}, \code{"xi"} or \code{"all"}. The name of the parameter whose
estimates are plotted. Use \code{"all"} to select all 3 GEV parameters.
}}

\item{distn}{A character scalar. The default is \code{"marginal"}.
\itemize{
\item \code{"marginal"}: 4 histograms of point estimates.
\itemize{
\item If \code{what = "return"} then 4 histograms of the point estimates of the
required return level are produced, one plot for each of the
approaches \code{"full"}, \code{"adjust"}, \code{"naive"} and \code{"discard"}.
\item If \code{what = "all"} then 3 histograms of the respective point
estimates of the GEV parameters \eqn{\mu}, \eqn{\sigma} and \eqn{\xi}
are produced, for the approach supplied in \code{approach}.
\item if \code{what = "mu"}, \code{"sigma"} or \code{"xi"} then 4 histograms of the
point estimates of the chosen GEV parameter are produced, one for
each of the approaches \code{"full"}, \code{"adjust"}, \code{"naive"} and
\code{"discard"}.
}
\item \code{"joint"}:
\itemize{
\item If \code{what = "return"} or \code{what = "mu"}, \code{"sigma"} or \code{"xi"} then
\code{\link[graphics:pairs]{pairs}} is used to produce a matrix of scatter plots
to compare the estimates produced by the approaches.
\item If \code{what = "all"} then 3 scatter plots are produced, one for each
of the GEV parameters \eqn{\mu}, \eqn{\sigma} and \eqn{\xi}, to
compare the estimates from the 2 approaches supplied in \code{approach}.
}
}}

\item{approach}{A character scalar or vector. A proper subset of
\code{c("full", "adjust", "naive", "discard")}. Only relevant if \code{what} is
\code{"all"}. If \code{distn = "marginal"} then the default is \code{"full"}.
If \code{distn = "joint"} then the default is \code{c("full", "adjust")}.}

\item{return_period}{Either a numeric scalar (for \code{plot.evmiss_sim_study_2})
or a numeric vector (for \code{summary.evmiss_sim_study_2} or
\code{tab.evmiss_sim_study_2}). A return period or periods, in numbers of blocks,
for which to compare the four estimation approaches. All values must be
greater than 1.}

\item{vertical}{logical scalar. Only relevant if \code{distn = "marginal"} or if
\code{distn = "joint"} and \code{what == "all"}. If \code{vertical = TRUE} then the plots
are arranged vertically. Otherwise, 4 plots are 2 rows of 2 plots and 3
plots are arranged horizontally.}

\item{main}{A character vector. Titles for the histograms.}

\item{penultimate}{A logical scalar. If  \code{distn = "marginal"} should
lines be added to plots to indicate the GEV parameter values, or the
implied return levels, from a penultimate approximation based on the
simulation distribution and the block length. See \code{\link[mev:smith.penult]{mev::smith.penult}}.}

\item{line_col}{A vector giving the colours of lines added to plots when
\code{distn = "marginal"}. \code{line_col[1]}: sample median of estimates;
\verb{line_col[2]:} sample mean of estimates; \code{line_col[3]}: true return level
(if \code{what = "return"}); \code{line_col[4]}: value from an penultimate
approximation (if \code{penultimate = TRUE}).}

\item{line_lty}{A numeric vector giving the line types of these vertical
lines.}

\item{legend}{A numeric vector of length 4. In which of the 4 plots should a
legend relating to the vertical lines be placed. The default is in the
final plot, the \code{"discard"} plot.}

\item{mar}{A numeric vector. The number of lines of margin on the four sides
of plots. See \code{\link[graphics:par]{par}}.}

\item{layout}{A logical scalar. Set \code{layout = TRUE} if
\code{\link[graphics:layout]{layout}} is used to create a multi-plot layout
external to the call to \code{plot.evmiss_sim_study_2}.}

\item{no_ylab}{A logical scalar. IF \code{no_ylab = TRUE} then no axis labels are
placed on the vertical axis.}

\item{...}{Further arguments. Only used in the following cases.
\itemize{
\item \code{plot.evmiss_sim_study_2}: to pass graphical parameters to the graphical
functions \code{\link[graphics:plot.default]{plot}}, \code{\link[graphics:hist]{hist}}.
\item \code{print.summary.evmiss_sim_study_2}: if \code{quiet = FALSE}, to pass arguments
to \code{\link[base:print]{print}}, such as \code{digits}.
\item \code{tab.evmiss_sim_study_2}: to pass arguments to
\code{summary.evmiss_sim_study_2}, such as \code{vsfull}.
}}

\item{statistics}{A function used to calculate statistics to summarise the
differences between estimates and the quantities described in \code{vsfull}
below. The object returned from this function should be a named vector.
See \strong{Examples}. If \code{statistics} is missing then the default is to
calculate the following sample statistics: mean, standard deviation, root
mean squared error, median, inter-quartile range, mean absolute error and
\code{isNA}, the number of simulations for which the GEV parameter MLEs are
missing owing to a convergence problem.}

\item{meanSE}{A logical scalar. If \code{meanSE = TRUE} and \code{vsfull = FALSE} then
the sample mean of the estimated standard errors over all simulated
datasets is added to the output summary statistics. Otherwise, this is not
added because a comparison of the sample standard deviation of an
estimator to \code{meanSE} is only valid for \code{vsfull = FALSE}.}

\item{log}{A logical scalar. If \code{log = TRUE} then make comparisons for the
GEV scale parameter \eqn{\sigma} based on estimate of \eqn{\log \sigma.}
Otherwise, based them on estimates of \eqn{\sigma}.}

\item{log_return}{A logical scalar. If \code{log_return = TRUE} then make
comparisons for return levels on the log scale. Otherwise, use the
original return level scale.}

\item{vsfull}{A logical scalar. If \code{vsfull = TRUE} then comparisons are made
relative to the estimates from the \code{"full"} case, that is, fitting to data
with no missing values. If \code{vsfull = FALSE} then comparisons are made
relative to the true values (\code{what = "return"}) or penultimate
approximations to the GEV parameter values (\code{what = "all"}).}

\item{quiet}{A logical scalar. If \code{quiet = TRUE} then do not print the
summary to the console.}
}
\value{
\strong{Plot method}. No return value, only the plots are produced.

\strong{Summary method}. A list object, with class
\code{c("summary.evmiss_sim_study_2", "list")}, containing the original function
call to \code{summary.evmiss_sim_study_2} and the following components, depending
on \code{what} and \code{vsfull}.

If \code{what = "all"}, matrices \code{full_gev}, \code{adjust_gev} and \code{naive_gev}.
Columns 1 to 3 contain the values of the summary statistics relating to
the estimates of \eqn{\mu}, \eqn{\sigma}, \eqn{\xi} respectively.

If \code{what = "return"}, matrices \code{full_rl}, \code{adjust_rl}, \code{naive_rl}.
The columns contain the values of the summary statistics for return levels
corresponding to the return periods in \code{return_period}.

The object is printed by \code{\link{print.summary.evmiss_sim_study_2}}. The rows
in the matrices described above are interleaved so that values for the
same summary statistic are printed in adjacent rows. If \code{vsfull = TRUE}
then all the values in \code{full_gev} and \code{full_rl} are zero and therefore are
not included when printing. If the elements of the vector returned from
the function \code{statistics} are named, then these names are included in the
row names. Otherwise, the names used are \code{stat1}, \code{stat2} etc.
}
\description{
Methods for objects of class \code{"evmiss_sim_study_2"} returned from
\code{\link{sim_study_2}}.
}
\details{
\strong{Plot method}. The S3 plot method provides plots to compare these 4
approaches based on the point estimates that they produce for return levels
and GEV parameters. Optionally, if \code{penultimate = TRUE}, comparisons can be
made with a penultimate GEV approximation to the true distribution of the
relevant block maxima.

\strong{The easiest way to appreciate the various plotting options that are
available is to look at the examples in} \code{\link{sim_study_2}}.

\strong{Summary method}. A matrix containing the values of the statistics in
\code{statistics} used to summarise the performance of estimators of GEV
parameters (if \code{what = "all"}) or return levels (if `what = "return"). The
columns relate to the GEV parameters or return levels. Each group of 4 rows
relates to a different combination of summary statistic and estimator
(approach to dealing with missingness), indicated by the row names.

\strong{Table method, tab}. Creates a table of results in which there is a row
for each estimator. If \code{vsfull = FALSE} is passed to
\code{summary.evmiss_sim_study_2} then the first row relates to the \code{"full"}
approach. Otherwise, only \code{"adjust"}, \code{"naive"} and \code{"discard"} are
included.
\itemize{
\item If \code{return_period} is supplied then the statistics concern
estimation of these return periods and the statistics included are
mean, standard deviation, root mean squared error, median, inter-quartile
range and mean absolute error.
\item Otherwise, the statistics concern the estimation of the GEV parameters and
the statistics included are mean, standard deviation and root mean squared
error.
}
}
\section{Examples}{
 See \code{\link{sim_study_2}}.
}

\seealso{
\code{\link{sim_study_2}}.
}
